pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Ciastelix", email = "tenlegitmati@gmail.com"},
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "dependency-injector>=4.41.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.16",
    "werkzeug>=2.3.6",
    "passlib>=1.7.4",
    "aiokafka>=0.8.1",
    "python-jose[cryptography]>=3.3.0",
    "requests>=2.31.0",
    "pydantic<2,>1.8",
    "bcrypt>=4.0.1",
    "pydantic-core>=2.3.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}
